fix warnings

+++ tests/cpp-tests/Classes/PerformanceTest/PerformanceContainerTest.cpp
@@ -156,7 +156,7 @@ void PerformanceContainerScene::initWithQuantityOfNodes(unsigned int nNodes)
     addChild(menuLayer);
     menuLayer->release();
 
-    log("Size of Node: %u\n", sizeof(Node));
+    log("Size of Node: %d\n", (int)sizeof(Node));
     
     int oldFontSize = MenuItemFont::getFontSize();
     MenuItemFont::setFontSize(24);
